# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Carga Pesada",
        "description": "API RESTFul Carga Pesada",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/login/access-token": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "operationId": "a3d8305ac0eb002633c8fe6aa3e3b08a",
                "requestBody": {
                    "description": "Dados de autenticação",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf",
                                    "password",
                                    "device_name"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678900"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "auth"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário autenticado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "token"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/Autenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The provided credentials are incorrect."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/login/check-token": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verifica se o token de acesso é válido",
                "operationId": "f04f47827fa518afab7a0104f6fbbd52",
                "responses": {
                    "200": {
                        "description": "Usuário autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido ou expirado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/login/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Deslogar o usuário",
                "operationId": "c15d14b0dd6d48885812c11fa686461b",
                "responses": {
                    "200": {
                        "description": "Logout bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout bem-sucedido."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Autenticated": {
                "required": [
                    "id",
                    "cpf",
                    "password",
                    "device_name"
                ],
                "properties": {
                    "id": {
                        "description": "ID único do contato (UUID)",
                        "type": "string",
                        "format": "uuid"
                    },
                    "cpf": {
                        "description": "CPF",
                        "type": "string"
                    },
                    "password": {
                        "description": "Password",
                        "type": "string"
                    },
                    "device_name": {
                        "description": "device_name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "required": [
                    "id",
                    "nome",
                    "email",
                    "cpf",
                    "rg",
                    "data_nascimento",
                    "ativo",
                    "superUser",
                    "permissions"
                ],
                "properties": {
                    "id": {
                        "description": "ID único do usuário",
                        "type": "string",
                        "format": "uuid"
                    },
                    "nome": {
                        "description": "Nome completo do usuário",
                        "type": "string"
                    },
                    "email": {
                        "description": "Endereço de e-mail do usuário",
                        "type": "string",
                        "format": "email"
                    },
                    "cpf": {
                        "description": "Número do CPF do usuário",
                        "type": "string"
                    },
                    "rg": {
                        "description": "Número do RG do usuário",
                        "type": "string"
                    },
                    "data_nascimento": {
                        "description": "Data de nascimento do usuário (no formato YYYY-MM-DD)",
                        "type": "string",
                        "format": "date"
                    },
                    "ativo": {
                        "description": "Indica se o usuário está ativo ou não",
                        "type": "boolean"
                    },
                    "superUser": {
                        "description": "Indica se o usuário possui privilégios de super usuário",
                        "type": "boolean"
                    },
                    "permissions": {
                        "description": "Lista de permissões do usuário",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Permission"
                        }
                    }
                },
                "type": "object"
            },
            "Permission": {
                "title": "Permission",
                "required": [
                    "id",
                    "nome",
                    "descricao"
                ],
                "properties": {
                    "id": {
                        "description": "ID único da permissão",
                        "type": "string",
                        "format": "uuid"
                    },
                    "nome": {
                        "description": "Nome da permissão",
                        "type": "string"
                    },
                    "descricao": {
                        "description": "Descrição da permissão",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}